+++ src/main/java/app/repository/MessageRepository.java
@@ -7,8 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface MessageRepository extends JpaRepository<Message, Integer> {
+public interface MessageRepository extends JpaRepository<Message, Long> {
 
-    // Get all messages for a conversation
-    List<Message> findByConversationIdOrderByCreatedAtAsc(int conversationId);
+    List<Message> findByConversationIdOrderByCreatedAtAsc(Long conversationId);
 }
